Lab4/ark-pzpi-23-4-fedkova-kateryna-lab4/caffinity/requirements.txt
asgiref==3.10.0
attrs==25.4.0
cffi==2.0.0
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.8
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg==3.2.13
psycopg-binary==3.2.13
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
